Fix for nested types being externally visible.  CA1034 https://docs.microsoft.com/en-us/visualstudio/code-quality/ca1034?view=vs-2019
