+++ src/modules/launcher/Plugins/Microsoft.Plugin.Program/Storage/PackageRepository.cs
@@ -14,12 +14,12 @@ namespace Microsoft.Plugin.Program.Storage
     /// A repository for storing packaged applications such as UWP apps or appx packaged desktop apps.
     /// This repository will also monitor for changes to the PackageCatelog and update the repository accordingly
     /// </summary>
-    internal class PackageRepository : ListRepository<UWP.Application>, IProgramRepository
+    internal class PackageRepository : ListRepository<UWPApplication>, IProgramRepository
     {
-        private IStorage<IList<UWP.Application>> _storage;
+        private IStorage<IList<UWPApplication>> _storage;
 
         private IPackageCatalog _packageCatalog;
-        public PackageRepository(IPackageCatalog packageCatalog, IStorage<IList<UWP.Application>> storage)
+        public PackageRepository(IPackageCatalog packageCatalog, IStorage<IList<UWPApplication>> storage)
         {
             _storage = storage ?? throw new ArgumentNullException("storage", "StorageRepository requires an initialized storage interface");
             _packageCatalog = packageCatalog ?? throw new ArgumentNullException("packageCatalog", "PackageRepository expects an interface to be able to subscribe to package events");
@@ -76,7 +76,7 @@ namespace Microsoft.Plugin.Program.Storage
             var windows10 = new Version(10, 0);
             var support = Environment.OSVersion.Version.Major >= windows10.Major;
 
-            var applications = support ? Programs.UWP.All() : new Programs.UWP.Application[] { };
+            var applications = support ? Programs.UWP.All() : new Programs.UWPApplication[] { };
             Set(applications);
         }
 
@@ -87,7 +87,7 @@ namespace Microsoft.Plugin.Program.Storage
 
         public void Load()
         {
-            var items = _storage.TryLoad(new Programs.UWP.Application[] { });
+            var items = _storage.TryLoad(new Programs.UWPApplication[] { });
             Set(items);
         }
     }
